Update dependencies and actions
This commit updates the versions of dependencies and actions in the python package workflow. Specifically, it updates the checkout action from v2 to v3 and the setup-python action from v2 to v3. It also updates the version of the short-sha action from v1.2 to v2.2, and the docker/build-push-action from v2 to v4.